Move ssao texture to TextureViewResources
